feat(front) reestrutura componentes das pages
