+++ front/src/pages/CadastroPage/components/CadastroForm/index.tsx
@@ -11,9 +11,9 @@ import {
 import { useForm } from '@mantine/form';
 import { showNotification } from '@mantine/notifications';
 import { Link, useNavigate } from 'react-router-dom';
+import { authCadastro } from '../../../../services/auth';
+import { notify, TypeNotificationEnum } from '../../../../utils/notify';
 import logo from '../../assets/logo.svg';
-import { authCadastro } from '../../services/auth';
-import { notify, TypeNotificationEnum } from '../../utils/notify';
 import { useStyles } from './styles';
 
 export function CadastroForm() {
