+++ front/src/pages/CustomPage/components/CategoriaList/components/EditCategoriaModal/index.tsx
@@ -16,33 +16,66 @@ import {
 } from '@mantine/core';
 import { useForm } from '@mantine/form';
 import { showNotification } from '@mantine/notifications';
-import { useContext, useState } from 'react';
-import { BiCategory, BiCustomize } from 'react-icons/bi';
-import AuthContext from '../../../../context/AuthContext/AuthContext';
-import { createCategoria } from '../../../../services/categoria';
-import { queryClient } from '../../../../services/queryClient';
-import { SelectItemIcon } from '../../../../utils/customSelect';
-import { notify, TypeNotificationEnum } from '../../../../utils/notify';
+import { useContext, useEffect, useState } from 'react';
+import { BiCategory } from 'react-icons/bi';
+import AuthContext from '../../../../../../context/AuthContext/AuthContext';
+import { useModalController } from '../../../../../../context/ModalContext/ModalContext';
 import {
-  getCategoriaIcon,
-  getSimpleIcon,
-  iconList,
-  StatusEnum,
-} from '../../../../utils/constants';
+  CategoriaResponse,
+  updateCategoria,
+} from '../../../../../../services/categoria';
+import { queryClient } from '../../../../../../services/queryClient';
+import { getSimpleIcon, iconList } from '../../../../../../utils/constants';
+import { SelectItemIcon } from '../../../../../../utils/customSelect';
+import { notify, TypeNotificationEnum } from '../../../../../../utils/notify';
+
 import { useStyles } from '../styles';
 
-interface CreateCategoriaModalProps {
+interface EditCategoriaModalProps {
   isOpen: boolean;
   onClose: () => void;
+  categoriaList: CategoriaResponse[];
 }
 
-export function CreateCategoriaModal({
+export function EditCategoriaModal({
   isOpen,
   onClose,
-}: CreateCategoriaModalProps) {
+  categoriaList,
+}: EditCategoriaModalProps) {
   const { classes } = useStyles();
   const { token } = useContext(AuthContext);
   const [Loading, setloading] = useState(false);
+
+  const { id } = useModalController();
+
+  const form = useForm({
+    initialValues: {
+      nome: '',
+      icon: 'CUSTOM',
+      color: '#6b6b6b',
+      isForReceita: true,
+      isForDespesa: true,
+    },
+    validate: (values) => ({
+      nome: values.nome === '' ? 'nome é obrigatório' : null,
+    }),
+  });
+
+  useEffect(() => {
+    const categoriaToEdit = categoriaList.find(
+      (categoria) => categoria.id === id
+    );
+
+    if (categoriaToEdit) {
+      form.setValues({
+        nome: categoriaToEdit.nome,
+        color: categoriaToEdit.color,
+        icon: categoriaToEdit.icon,
+        isForDespesa: categoriaToEdit.isForDespesa,
+        isForReceita: categoriaToEdit.isForReceita,
+      });
+    }
+  }, [isOpen]);
 
   const handleClose = () => {
     form.reset();
@@ -58,7 +91,7 @@ export function CreateCategoriaModal({
         })
       );
     } else {
-      createCategoria(data, token.token)
+      updateCategoria(id, data, token.token)
         .then(() => {
           queryClient.invalidateQueries('categorias').then(() => {
             showNotification(notify({ type: TypeNotificationEnum.SUCCESS }));
@@ -80,19 +113,6 @@ export function CreateCategoriaModal({
     }
   };
 
-  const form = useForm({
-    initialValues: {
-      nome: '',
-      icon: '',
-      color: '#6b6b6b',
-      isForReceita: true,
-      isForDespesa: true,
-    },
-    validate: (values) => ({
-      nome: values.nome === '' ? 'nome é obrigatório' : null,
-    }),
-  });
-
   return (
     <Modal
       opened={isOpen}
@@ -105,7 +125,7 @@ export function CreateCategoriaModal({
           <ActionIcon color="blue" variant="outline" size={40}>
             <BiCategory size={40} />
           </ActionIcon>
-          <Title order={3}>Cadastro de Categoria</Title>
+          <Title order={3}>Edição de Categoria</Title>
         </Box>
       }
     >
