+++ front/src/pages/CustomPage/components/CategoriaList/components/styles.ts
