+++ front/src/pages/Home/components/InfoCards/Card/index.tsx
