+++ front/src/pages/Home/components/InfoCards/Card/styles.ts
