+++ front/src/pages/Home/components/TransacaoList/TransacaoItem/styles.ts
