+++ front/src/pages/Home/index.tsx
@@ -1,9 +1,6 @@
-import { useContext, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { Navigation } from '../../components/Navigation';
-import { MainDashboard } from '../../components/MainDashboard';
-import AuthContext from '../../context/AuthContext/AuthContext';
 import { MonthProvider } from '../../context/MonthContext/MonthContext';
+import { MainDashboard } from './components/MainDashboard';
 
 export function Home() {
   return (
